+++ tablets/delete_issue_tablets.php
@@ -1,8 +1,16 @@
 <?php
+session_start();
+if($_SESSION['user_id'] == session_id())
+{
 	$connection = mysqli_connect("remotemysql.com","bjBEokAclb","q8e75oGyd0");
 	$db = mysqli_select_db($connection,"bjBEokAclb");
 	$query = "delete from t_issue where s_id = '$_GET[id]'";
 	$query_run = mysqli_query($connection,$query);
+}
+else
+	{
+		header("location:../index.php");
+	} 
 ?>
 <script type="text/javascript">
 	alert("Student Deleted...");
